Fix for empty field in WishModel
On wish creation, the WishModel expects a suggested_for_user_id to be a UUID or None.
The frontend sends an empty string, which creates an error as it is neither UUID or None.

Fix applied to force suggested_for_user_id to be None if this is an empty string.